removed unwanted packages replaced with directories, reduced over engineering
