+++ app/domain/dependencies/blink_detector.py
