+++ app/domain/dependencies/face_mesh_detector.py
