+++ app/domain/dependencies/gaze_detector.py
