+++ app/domain/detectors/facial_cue_detector.py
@@ -2,14 +2,14 @@ import cv2
 import os
 import time
 
-from app.infrastructure.capture.detectors.blink_detector import extract_eye_landmarks, calculate_ear, draw_eye_outline, detect_blinks
-from app.infrastructure.capture.detectors.face_expression_detector import expression_counts, detect_expression
-from app.infrastructure.capture.detectors.face_mesh_detector import FaceMeshDetector
-from app.infrastructure.capture.detectors.gaze_detector import extract_iris_center, extract_iris_landmarks, analyze_gaze_for_eye, \
-    draw_iris_outline, draw_iris_center
-from app.infrastructure.capture.detectors.yawn_detector import extract_mouth_landmarks, calculate_mar, draw_mouth_state, detect_yawn
+from app.domain.dependencies.face_mesh_detector import FaceMeshDetector
+from app.domain.dependencies.blink_detector import extract_eye_landmarks, calculate_ear, draw_eye_outline, detect_blinks
+from app.domain.dependencies.face_expression_detector import detect_expression
+from app.domain.dependencies.gaze_detector import (extract_iris_center, extract_iris_landmarks, analyze_gaze_for_eye,
+                                                   draw_iris_outline, draw_iris_center)
+from app.domain.dependencies.yawn_detector import extract_mouth_landmarks, calculate_mar, draw_mouth_state, detect_yawn
 
-class FacialCueAnalyzer:
+class FacialCueDetector:
     def __init__(self):
         # Counters (previously module-level globals)
         self.blink_counts = 0
@@ -27,13 +27,6 @@ class FacialCueAnalyzer:
 
         # Detector instance (was module-level)
         self.face_detector = FaceMeshDetector()
-        # Camera handle stored on instance for coordinated cleanup
-        self.capture = None
-        # Headless/GUI control via env (default to headless for API)
-        self.show_gui = os.getenv("CLD_SHOW_GUI", "0").lower() in ("1", "true", "yes")
-        # Periodic capture refresh to avoid long-running freezes
-        self.refresh_interval_sec = int(os.getenv("CLD_CAPTURE_REFRESH_SEC", "60"))
-        self._last_refresh = time.time()
 
         self.facial_cues_data = {
             "blink_counts": 0,
@@ -62,25 +55,12 @@ class FacialCueAnalyzer:
 
     def start_facial_cue_detector(self):
         self.running = True
-        # Prefer DirectShow on Windows for faster open/close
-        try:
-            self.capture = cv2.VideoCapture(0, cv2.CAP_DSHOW)
-        except Exception:
-            self.capture = cv2.VideoCapture(0)
-        # Reduce internal buffering where supported
-        try:
-            self.capture.set(cv2.CAP_PROP_BUFFERSIZE, 1)
-        except Exception:
-            pass
+        capture = cv2.VideoCapture(0)
 
         while self.running:
-            ret, frame = self.capture.read() if self.capture is not None else (False, None)
+            ret, frame = capture.read()
             if not ret:
                 print("Can't receive frame (stream end?). Exiting ...")
-                break
-
-            # Quick exit path if stop requested to reduce latency
-            if not self.running:
                 break
 
             frame, face_landmarks = self.face_detector.detect_face_landmarks(frame)
@@ -176,38 +156,14 @@ class FacialCueAnalyzer:
                     if last_expression in self.facial_cues_data["face_expression_counts"]:
                         self.facial_cues_data["face_expression_counts"][last_expression] += 1
 
-            if self.show_gui:
-                cv2.imshow("Facial Cue Detection", frame)
-                if cv2.waitKey(1) & 0xFF == ord('q'):
-                    self.stop_facial_cue_detector()
-                    break
+            cv2.imshow("Facial Cue Detection", frame)
 
-            # Periodic capture refresh to avoid freezing after long runs
-            if self.refresh_interval_sec > 0 and (time.time() - self._last_refresh) >= self.refresh_interval_sec:
-                # Re-open the camera quickly; keep processing loop alive
-                if self.capture is not None:
-                    try:
-                        self.capture.release()
-                    except Exception:
-                        pass
-                try:
-                    self.capture = cv2.VideoCapture(0, cv2.CAP_DSHOW)
-                except Exception:
-                    self.capture = cv2.VideoCapture(0)
-                try:
-                    self.capture.set(cv2.CAP_PROP_BUFFERSIZE, 1)
-                except Exception:
-                    pass
-                self._last_refresh = time.time()
+            if cv2.waitKey(1) & 0xFF == ord('q'):
+                self.stop_facial_cue_detector()
+                break
 
-        # Cleanup performed in the same thread that created the resources
-        if self.capture is not None:
-            try:
-                self.capture.release()
-            finally:
-                self.capture = None
-        if self.show_gui:
-            cv2.destroyAllWindows()
+        capture.release()
+        cv2.destroyAllWindows()
         self.face_detector.close()
 
     def facial_cue_snap_shot_and_reset(self):
@@ -238,11 +194,7 @@ class FacialCueAnalyzer:
         }
 
     def stop_facial_cue_detector(self):
-        # Only signal stop and reset data. Cleanup happens in the worker thread
-        # to avoid cross-thread OpenCV GUI operations which can freeze on Windows.
         self.running = False
         self.reset_data()
-
-
-
-
+        cv2.destroyAllWindows()
+        self.face_detector.close()
