+++ app/infrastructure/database/ConfigureDatabase.py
@@ -2,7 +2,6 @@ import os
 from pymongo import MongoClient
 from pymongo.errors import ServerSelectionTimeoutError
 
-
 class ConfigureDatabase:
     def __init__(self,
                  uri=None,
@@ -10,7 +9,6 @@ class ConfigureDatabase:
         self.client = None
         self.database = None
         try:
-            # Read from env or explicit params. No hardcoded credentials.
             uri = uri or os.getenv("MONGODB_URI")
             database_name = database_name or os.getenv("MONGODB_DB")
